pyproject.toml
[project]
name = "fact-shield"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.8",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "langgraph-cli[inmem]>=0.3.6",
    "langsmith>=0.4.8",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "streamlit>=1.39.0",
    "tavily-python>=0.7.11",
]
